Refactor app state handling
Switch from using an indexed state monad to more straightforward
functions that accept and return an AppState. While the state monad
approach was educational, it started to seem like an over-engineered
solution.

Event handling is now done by constructing a pipeline of functions
via regular or Kleisli composition that the AppState is sent through.